Move CLI-specific code to a separate module
